requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6

# ML and embeddings
openai>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
faiss-cpu>=1.7.0

# Data processing
python-docx>=1.1.0
PyPDF2>=3.0.0
aiofiles>=23.0.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Visualization
plotly>=5.15.0

# Optional for production
redis>=5.0.0
celery>=5.3.0

# Development
pytest>=7.4.0
black>=23.11.0
flake8>=6.1.0
